refactor: move all dashboard main page components to /dashboard/main folder
