+++ src/app/dashboard/layout.tsx
@@ -1,5 +1,5 @@
 import Header from "@/components/header/Header";
-import SideBar from "@/components/dashboard/organisms/SideBar";
+import SideBar from "@/components/dashboard/main/organisms/SideBar";
 
 export default function DashboardLayout({
   children,
