+++ src/app/dashboard/page.tsx
@@ -1,13 +1,13 @@
-import DashboardOverview from "@/components/dashboard/molecules/Overview";
-import DashboardTransaction from "@/components/dashboard/molecules/Transaction";
-import DashboardCalendarShortcut from "@/components/dashboard/molecules/CalendarShortcut";
-import DashboardSchedule from "@/components/dashboard/molecules/Schedule";
-import DashboardUpcomingDue from "@/components/dashboard/molecules/UpcomingDue";
-import DashboardRanking from "@/components/dashboard/molecules/Ranking";
+import DashboardOverview from "@/components/dashboard/main/molecules/Overview";
+import DashboardTransaction from "@/components/dashboard/main/molecules/Transaction";
+import DashboardCalendarShortcut from "@/components/dashboard/main/molecules/CalendarShortcut";
+import DashboardSchedule from "@/components/dashboard/main/molecules/Schedule";
+import DashboardUpcomingDue from "@/components/dashboard/main/molecules/UpcomingDue";
+import DashboardRanking from "@/components/dashboard/main/molecules/Ranking";
 
 export default function DashboardPage() {
   return (
-    <div className="grid grid-cols-[minmax(0,_1fr)_minmax(0,_1fr)_minmax(0,_1fr)_minmax(0,_390px)] auto-rows-min gap-3 w-full p-[38px]">
+    <section className="grid grid-cols-[minmax(0,_1fr)_minmax(0,_1fr)_minmax(0,_1fr)_minmax(0,_390px)] auto-rows-min gap-3 w-full p-[38px]">
       <DashboardOverview />
       <div className="col-span-3 flex flex-col gap-3">
         <DashboardTransaction />
@@ -18,6 +18,6 @@ export default function DashboardPage() {
         <DashboardUpcomingDue />
         <DashboardRanking />
       </div>
-    </div>
+    </section>
   );
 }
